Viser skjemaet ved utfylling av tilleggsmottaker nedenfor første mottaker, slik at SB ikke mister den av syne.
